Ruby running animation, cog projectiles, Robot moving and fixed animations
Ruby is now able to look like she's running in different directions when she moves, so she's no longer stuck on one sprite. I also added  the cog projectile to Ruby and ensured that it doesn't collide with Ruby. The Robot enemy also has movement animations, and when hit by Ruby's cog projectile, they dance to signify they are fixed. Had to modify the Robot's collider to ensure it would be hit by the cog.